I didn't like the nondescript A and B from Pair, so I created a new Class WayTriplet.
It almost works the way I want it to, now. It should also split if the first or last way in one of the parents is reached in case the line currently processed continues on a stop that is a terminus for other lines.

An odd bug is that the way that ends up in previousWay in the first iteration, stays in the original relation.

One regression is that identical relations are created multiple times, instead of being created once and then reused in the superroute