In some cases the membershipcount is needed of all the (parent) route relations, in others, it's only the routes that go in the same sense that count.
All tests for 2 lines that make a circle around the city, but which at the railway station and the bus station near the hospital reuse ways in the same or the opposite sense, pass now.

Now I'm going to do some interactive testing on other bus lines.

I also found a bus line that wasn't mapped correctly, so the test osm file is also updated