In ExtractRelationMembersToNewRelationAction I made an attempt to get rid of the Index out of Bounds exception. The attempt failed.
Fixed some bugs in RouteSegmentToExtract that came from reversing the order in which members are evaluated

Tried to add functionality to the test code generator to copy to the clipboard. Unfortunately it doesn't work on Linux or on OpenJDK.

I had to change the OSM file. Some data that is important for the decision where to split was missing from it.