Fix #22233: Deadlock when uploading with incomplete member download enabled
This is caused by custom threading and blocking during the
validation process on upload.

The validation is run in the EDT, but the custom threading and blocking
created a new thread, started it, and then just blocked the EDT. The
new thread would then attempt to download data, and during that process
it would attempt to show a UI window. This is where the deadlock would
occur (the EDT thread is blocked on this new thread).

Signed-off-by: Taylor Smock <tsmock@fb.com>